Fixed some tests that were failing on arzach due to it using numpy-2.0dev.  First of all, there were some subtle differences in array string representation that were causing some tests to fail.  I fixed the tests to compare actual values instead of string representations.  And there were two tests failing due to numpy's now stricter typecasting on ufuncs.
git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@938 ed100bfc-0583-0410-97f2-c26b58777a21

+++ lib/pyfits/hdu/image.py
@@ -409,7 +409,7 @@ class _ImageBaseHDU(_ValidHDU):
                     # of adding because of the way numpy converts
                     # the raw signed array into an unsigned array.
                     data = np.array(data, dtype=dtype)
-                    data -= (1 << (bits - 1))
+                    data -= np.uint64(1 << (bits - 1))
                     return data
 
     def _dimShape(self):
